refactor(Wallet) fix naming and styling
